+++ ChangeLog.md
@@ -1,3 +1,11 @@
+# v13.5.0
+
+- BREAKING CHANGE: Disable use of _Perceptive_ as a visibility handler until such time as I can make it work
+- Restructure internal loops to improve batch processing of visibility changes
+- Added opt-in game setting to allow stealth initiative rolls to apply against allies.
+- Added opt-in game setting to disable writing stealth summaries to initiative cards.
+- Added opt-in advanced game setting to use Visioner bulk API
+
 # v13.4.1
 
 - Restrict clear party stealth keybinding to GM only